api/requirements.txt
# UI-Agnostic API Dependencies

fastapi>=0.110.0
uvicorn>=0.29.0
langgraph>=0.0.30
langchain-openai>=0.1.3
langchain-core>=0.1.45
pydantic>=2.4.0
httpx>=0.25.0
python-dotenv>=1.0.0
